# cmd/docsync: new command