--- a/bin/upload-docs
+++ b/bin/upload-docs
@@ -23,16 +23,18 @@ docgenerate $sourcePath $generate_dest_path
 
 echo
 echo "Uploading docs...."
-aws s3 sync --delete $generate_dest_path $bucket_url/docs
+docsync $bucket $generate_dest_path
 
-echo
-echo "Setting content type for extensionless files, this will take a while..."
-cd $generate_dest_path
-find . -type f ! -name "*.*" |
-  sed -e "s/^.\\///" |
-  xargs -I {} aws s3api copy-object \
-    --bucket $bucket \
-    --content-type "text/html" \
-    --copy-source $bucket/docs/{} \
-    --key docs/{} \
-    --metadata-directive "REPLACE" > /dev/null
+# aws s3 sync --delete $generate_dest_path $bucket_url/docs
+#
+# echo
+# echo "Setting content type for extensionless files, this will take a while..."
+# cd $generate_dest_path
+# find . -type f ! -name "*.*" |
+#   sed -e "s/^.\\///" |
+#   xargs -I {} aws s3api copy-object \
+#     --bucket $bucket \
+#     --content-type "text/html" \
+#     --copy-source $bucket/docs/{} \
+#     --key docs/{} \
+#     --metadata-directive "REPLACE" > /dev/null