Add support for zero-downtime doc deploys by uploading files one at a time 
with correct the content type, and removing files not present on the local 
system after the fact.