@jeffomatic 2017-06-06
Would you mind doing an informal benchmark of the new tool vs the old one on an upload to staging?