@dominic 2017-06-06
Unscientific coffeeshop benchmark:

```
OLD:
bin/upload-docs  196.55s user 54.15s system 96% cpu 4:18.92 total

NEW:
bin/upload-docs  144.91s user 44.27s system 55% cpu 5:42.38 total
```